Completando circuito en Redux.
